refactor: inline index symbol normalization into Motilal loader
Drop utils/index_symbol_mapping and move the NSE/BSE alias maps + cleanup
helpers directly into broker/motilal/database/master_contract_db.py. The
mapping is Motilal-specific (other brokers already feed clean strings),
so a shared util is unnecessary and risked coupling unrelated brokers.

Functionality preserved from the removed util:
- NSE: upper-case + strip whitespace, then alias lookup; unlisted symbols
  fall through after basic cleanup.
- BSE: alias lookup with upper-case + whitespace-collapsed keys so
  "BSE  CAPGOOD" and "BSE CAPGOOD" both resolve; unlisted BSE indices
  fall back to upper + strip ("BSE 1000" -> "BSE1000").
- NaN guard in the apply lambdas preserves missing rows instead of
  coercing them to the literal string "NAN".
